Tidy up websockets?
