+++ lute/net/src/websocket_common.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "lua.h"
+
+#include <cstddef>
+
+namespace net
+{
+
+struct WebSocketPayload
+{
+    const char* data = nullptr;
+    size_t length = 0;
+    bool binary = false;
+};
+
+WebSocketPayload extractWebSocketPayload(lua_State* L, int index);
+
+} // namespace net
